backend/requirements.txt
flask
peewee
PyMySQL
pyyaml
Flask-Cors
flask-login
